Cleaned up zeset references and compiler warnings.
git-svn-id: svn+ssh://frodo/cluster/svn/projects@2288 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ psc_fsutil_libs/lnet-lite/lnet/config.c
@@ -1177,7 +1177,7 @@ lnet_match_networks (char **networksp, char *ip2nets, __u32 *ipaddrs, int nip)
                                         "%s%s", (len == 0) ? "" : ",", 
                                         tb->ltb_text);
                 
-                        if (len >= sizeof(networks)) {
+                        if (len >= (int)sizeof(networks)) {
                                 CERROR("Too many matched networks\n");
                                 rc = -E2BIG;
                                 goto out;
