+++ psc_fsutil_libs/lnet-lite/lnet/lib-me.c
@@ -38,7 +38,7 @@ LNetMEAttach(unsigned int portal,
         LASSERT (the_lnet.ln_init);
         LASSERT (the_lnet.ln_refcount > 0);
         
-        if (portal >= the_lnet.ln_nportals)
+        if (portal >= (unsigned int)the_lnet.ln_nportals)
                 return -EINVAL;
 
         me = lnet_me_alloc();
